nodes updated

+++ nodes.py
@@ -1,31 +1,35 @@
-import pandas as pd
-import numpy as np
-import os
-import geopandas as gpd
-
 """
 Nodes are derived from Navstreet Zlevel file
 Steps: Zlevel file > Drop nodeids with 0 > Combined 5 files into 1 node file for entire CA
 Population density extracted using ArcGIS is formated
 """
 
-zlevelpath = '../from_Here/original_HERE'
-filenames = ['here_map_11/here_map_11_shapefiles/Zlevels.shp', 'here_map_12-20191207T191053Z-001/here_map_12/here_map_12_shapefiles/Zlevels.shp',
-'here_map_13/here_map_13_shapefiles/Zlevels.shp','here_map_14/here_map_14_shapefiles/Zlevels.shp','here_map_15-20191207T193135Z-001/here_map_15/here_map_15_shapefiles/Zlevels.shp' ]
+import pandas as pd
+import numpy as np
+import os
+import geopandas as gpd
 
-print("=== Loading the Zlevel geodataframes =====")
-zleveldfs = []
-for file in filenames:
-    zleveldfs = [gpd.read_file(os.path.join(zlevelpath, file))]
-print("--- Number of dataframes are ---", len(zleveldfs))
-print("Length of first df", len(zleveldfs[0]))
-zlevel_combined = pd.concat([zleveldfs[i] for i in range(len(zleveldfs))])
-zlevel_combined.to_csv('../midstages/zlevel_combined.csv',index = False)
-print("=== Completed writing =====")
+if not os.path.isfile('../midstages/zlevel_combined.csv'):
+    print('====== Started combining ZLevel files ======')
+    zlevelpath = '../from_Here/original_HERE'
+    filenames = ['here_map_11/here_map_11_shapefiles/Zlevels.shp', 'here_map_12-20191207T191053Z-001/here_map_12/here_map_12_shapefiles/Zlevels.shp',
+    'here_map_13/here_map_13_shapefiles/Zlevels.shp','here_map_14/here_map_14_shapefiles/Zlevels.shp','here_map_15-20191207T193135Z-001/here_map_15/here_map_15_shapefiles/Zlevels.shp' ]
 
-#For Nodes files from z level(dropping 0 node ids)
-nodes = zlevel_combined[zlevel_combined.NODE_ID!=0] #6.5Million nodes
-print("Nodes number", len(nodes))
+    print("=== Loading the Zlevel geodataframes =====")
+    zleveldfs = []
+    for file in filenames:
+        zleveldfs = [gpd.read_file(os.path.join(zlevelpath, file))]
+    print("--- Number of dataframes are ---", len(zleveldfs))
+    print("Length of first df", len(zleveldfs[0]))
+    zlevel_combined = pd.concat([zleveldfs[i] for i in range(len(zleveldfs))])
+    zlevel_combined.to_csv('../midstages/zlevel_combined.csv',index = False)
+    print("=== Completed writing =====")
+
+print('====== Zlevel Combined already exists. Started  creating Nodes file ======')
+zlevel_combined = pd.read_csv('../midstages/zlevel_combined.csv') 
+print("Length of zlevel file", len(zlevel_combined))
+nodes = zlevel_combined[zlevel_combined.NODE_ID!=0] #For Nodes files from z level(dropping 0 node ids)
+print("Number of nodes", len(nodes))
 #changing the order of columns  #6.5Million nodes
 cols = nodes.columns.tolist()
 cols = cols[2:7] + [cols[1]] + [cols[7]] + [cols[0]]
@@ -40,11 +44,11 @@ print(nodes['Z_LEVEL'].value_counts())
 print(len(nodes.drop_duplicates(subset='geometry', keep="first"))) # length ofnode ids that are duplicated in geometry. This shows that some intersections with same geometry having differnt z levels are given differnt node ids.
 
 #Write to csv and shapefile
-nodes.to_csv('../midtages/nodes_all.csv', index = False)
-nodes.to_file(driver='ESRI Shapefile', filename='nodes_all.shp')
+nodes.to_csv('../midstages/nodes_all.csv', index = False)
+#nodes.to_file(driver='ESRI Shapefile', filename='nodes_all.shp')
 
 # Calculating Population density for nodes - File extracted using ArcGIS
 nodes_pop = pd.read_csv('../midtages/allnodes_popden.csv')
 nodes_pop.rename(columns = {'RASTERVALU': 'pop_density'}, inplace = True)
 nodes_pop = nodes_pop.drop(['OBJECTID'], axis = 1)
-nodes_pop.to_csv('../final/allnodes_population_density.csv', index = False)
+nodes_pop.to_csv('../midstages/allnodes_population_density.csv', index = False)
